requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
pymysql==1.1.0
python-dotenv==1.0.0
WTForms==3.1.2